+++ src/main/java/com/purej/vminspect/http/HttpRequest.java
@@ -1,6 +1,7 @@
 // Copyright (c), 2013, adopus consulting GmbH Switzerland, all rights reserved.
 package com.purej.vminspect.http;
 
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 import com.purej.vminspect.util.Utils;
@@ -22,8 +23,11 @@ public class HttpRequest {
     this.request = null;
   }
 
-  public HttpRequest(HttpServletRequest request) {
+  public HttpRequest(HttpServletRequest request) throws IOException {
     this.request = request;
+    // Set character-encoding before reading parameters. UTF-8 is required
+    // for application/x-www-form-urlencoded params sent in body.
+    this.request.setCharacterEncoding("UTF-8");
     // Add all parameters:
     for (var e = request.getParameterNames(); e.hasMoreElements();) {
       var name = e.nextElement();
@@ -49,6 +53,29 @@ public class HttpRequest {
    */
   public String getParameter(String name) {
     return parameters.get(name);
+  }
+
+  /**
+   * Returns the named parameter value for save-operations. This method ensure its
+   * a POST request and the parameter is from the body, exists and not shadowed with
+   * a fake URL parameter.
+   */
+  public String getSaveParameter(String name) {
+    if (request == null) {
+      return getParameter(name); // No servlet-request
+    }
+    if (!request.getMethod().equals("POST")) {
+      throw new SecurityException("POST required");
+    }
+    var query = request.getQueryString();
+    if (query != null && query.contains(name)) {
+      throw new SecurityException("Body parameter expected");
+    }
+    var result = parameters.get(name);
+    if (result == null) {
+      throw new SecurityException("Parameter missing");
+    }
+    return result;
   }
 
   /**